Изменяет структуру кода в модуле
